requirements.txt
crewai
openai
feedparser
python-dotenv